#6 계정 만들기 - 의존성 정리
issue : https://github.com/JustBurrow/kobalttown/issues/6
time : 13h 40m
