+++ account-jpa/build.gradle
@@ -3,7 +3,6 @@ dependencies {
   compile project(':support-spring-jpa')
 
   testCompile project(path: ':account-domain', configuration: 'testArchives')
-  testCompile project(':configuration-jpa')
 
   testCompile 'com.h2database:h2'
 }
