+++ account-service/build.gradle
@@ -1,8 +1,8 @@
 dependencies {
-  compile project(':account-domain')
+  implementation project(':account-domain')
   compile project(':account-dao')
   compile project(':configuration-bean')
-  implementation project(':configuration-security')
+  compile project(':configuration-security')
 
   compile project(':support-spring-boot')
