--- common-util/src/main/java/kr/lul/kobalttown/common/util/SecretHashEncoder.java
@@ -1,11 +1,0 @@
-package kr.lul.kobalttown.common.util;
-
-/**
- * @author justburrow
- * @since 2019-02-28
- */
-public interface SecretHashEncoder {
-  String encode(CharSequence plain);
-
-  boolean matches(CharSequence plain, String hash);
-}
