+++ configuration-security/build.gradle
@@ -1,7 +1,8 @@
 dependencies {
   implementation project(':common-util')
-  implementation project(':support-spring-security')
+  implementation project(':account-domain')
+  compile project(':account-jpa')
 
-  implementation 'org.springframework.security:spring-security-core'
-  implementation 'org.springframework:spring-core'
+  compile project(':support-spring-security')
+  compile project(':configuration-jpa')
 }
