+++ db/kobalttown.sql
@@ -1,5 +1,5 @@
 -- MySQL Script generated by MySQL Workbench
--- Sun Mar  3 19:23:26 2019
+-- Mon Mar  4 23:43:52 2019
 -- Model: New Model    Version: 1.0
 -- MySQL Workbench Forward Engineering
 
@@ -89,6 +89,28 @@ CREATE TABLE IF NOT EXISTS `kobalttown`.`log_account` (
   PRIMARY KEY (`lid`))
 ENGINE = InnoDB;
 
+SHOW WARNINGS;
+USE `kobalttown` ;
+
+-- -----------------------------------------------------
+-- View `kobalttown`.`view_account_credential`
+-- -----------------------------------------------------
+USE `kobalttown`;
+CREATE  OR REPLACE VIEW `view_account_credential` AS
+    SELECT 
+        `a`.*,
+        `c`.`id` AS `cid`,
+        `c`.`public_key`,
+        `c`.`secret_hash`,
+        `c`.`created_at` AS `credential_created_at`,
+        `c`.`modify_ts` AS `creadential_modify_ts`
+    FROM
+        `user_account` AS `a`,
+        `user_credential` AS `c`
+    WHERE
+        `a`.`id` = `c`.`account`
+    ORDER BY `a`.`id` ASC , `c`.`id` ASC
+;
 SHOW WARNINGS;
 USE `kobalttown`;
