+++ front/templates/page/_/index-user.html
@@ -21,7 +21,7 @@
         <button form="search">검색</button>
     </div>
     <div>
-        <a href="index-user.html">맨처음</a>
+        <a href="index-user.html" th:href="@{/}">맨처음</a>
         <ol start="10">
             <li><a href="index-user.html?page=10">이전 블록</a></li>
             <li><a href="index-user.html?page=11">11</a></li>
@@ -126,7 +126,7 @@
         </tbody>
     </table>
     <div>
-        <a href="index-user.html">맨처음</a>
+        <a href="index-user.html" th:href="@{/}">맨처음</a>
         <ol start="10">
             <li><a href="index-user.html?page=10">이전 블록</a></li>
             <li><a href="index-user.html?page=11">11</a></li>
