+++ support-spring-jpa/src/main/java/kr/lul/kobalttown/support/spring/jpa/SupportSpringJpaAnchor.java
@@ -1,0 +1,16 @@
+package kr.lul.kobalttown.support.spring.jpa;
+
+import kr.lul.kobalttown.common.util.Anchor;
+
+/**
+ * @author justburrow
+ * @since 2019-03-04
+ */
+public final class SupportSpringJpaAnchor implements Anchor {
+  public static final Package PACKAGE = SupportSpringJpaAnchor.class.getPackage();
+  public static final String PACKAGE_NAME = PACKAGE.getName();
+
+  public SupportSpringJpaAnchor() {
+    throw new UnsupportedOperationException();
+  }
+}
