+++ support-spring-web/build.gradle
@@ -1,4 +1,6 @@
 dependencies {
+  implementation project(':common-util')
+
   compile 'org.springframework.boot:spring-boot-starter-web'
   compile 'org.springframework.boot:spring-boot-starter-thymeleaf'
 }
